creating forms with urls known beforehand.
if the email is registered for the user,
the form is automatically confirmed. if not,
a confirmation email is sent automatically.
